requirements.txt
bidict
certifi
cffi
charset-normalizer
click
colorama
dnspython
eventlet
Flask
flask-ipban
Flask-ReCaptcha
Flask-SocketIO
gevent
greenlet
idna
importlib-metadata
itsdangerous
Jinja2
mariadb==1.0.11
MarkupSafe
Pillow
pycparser
python-engineio
python-socketio
PyYAML
qrcode
requests
six
urllib3
Werkzeug
zipp
zope.event
zope.interface
